RoboCar.PWM (robocar v0.1.0)

Control the hardware PWM via the sysfs interface

Functions

duty_cycle_path(pwm)

Specs

duty_cycle_path(%RoboCar.PWM{channel_num: integer(), chip_num: integer()}) ::
  String

Return the path to the sysfs pwm duty_cycle file for the given PWM struct

Examples

iex> RoboCar.PWM.new |> RoboCar.PWM.duty_cycle_path
"/tmp/sys/class/pwm/pwmchip0/pwm0/duty_cycle"

enable(pwm)

Specs

enable(%RoboCar.PWM{channel_num: integer(), chip_num: integer()}) ::
  :ok | {:error, term()}

Enable PWM channel

Examples

iex> RoboCar.PWM.new |> RoboCar.PWM.enable
:ok

enable_path(pwm)

Specs

enable_path(%RoboCar.PWM{channel_num: integer(), chip_num: integer()}) :: String

Return the path to the sysfs pwm enable file for the given PWM struct

Examples

iex> RoboCar.PWM.new |> RoboCar.PWM.enable_path
"/tmp/sys/class/pwm/pwmchip0/pwm0/enable"

new()

Specs

new() :: %RoboCar.PWM{channel_num: integer(), chip_num: integer()}

Create a new %RoboCar.PWM{} struct with default values

Examples

iex> RoboCar.PWM.new
%RoboCar.PWM{channel_num: 0, chip_num: 0}

new(chip_num, channel_num)

Specs

new(integer(), integer()) :: %RoboCar.PWM{
  channel_num: integer(),
  chip_num: integer()
}

Create a new %RoboCar.PWM{chip_num: chip_num, channel_num: channel_num} struct

Examples

iex> RoboCar.PWM.new(1, 1)
%RoboCar.PWM{channel_num: 1, chip_num: 1}

period_path(pwm)

Specs

period_path(%RoboCar.PWM{channel_num: integer(), chip_num: integer()}) :: String

Return the path to the sysfs pwm period file for the given PWM struct

Examples

iex> RoboCar.PWM.new |> RoboCar.PWM.period_path
"/tmp/sys/class/pwm/pwmchip0/pwm0/period"

sysfs_path(pwm)

Specs

sysfs_path(%RoboCar.PWM{channel_num: integer(), chip_num: integer()}) :: String

Return the path to the sysfs directory for the given PWM struct

Examples

iex> RoboCar.PWM.new |> RoboCar.PWM.sysfs_path
"/tmp/sys/class/pwm/pwmchip0/pwm0"